+++ src/core/range/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.?? (2021-??-??)
+
+#### :rocket: New Feature
+
+* Added the support of non-including range bounds
+
 ## v3.30.1 (2021-01-11)
 
 #### :bug: Bug Fix